\docType{data}
\name{albanese}
\alias{albanese}
\title{Correlation matrix for Albanese, Stasio, Chiacchio, Fiorilli, and Pons (2010)}
\format{A data frame with 5 observations on the following 5 variables.
\describe{ \item{list("Age")}{a numeric vector} \item{list("CPM")}{a numeric
vector} \item{list("External")}{a numeric vector} \item{list("Mental")}{a
numeric vector} \item{list("Reflexive")}{a numeric vector} }}
\source{
Albanese, O., Stasio, S., Chiacchio, C., Fiorilli, C., &
Pons, F. (2010). Emotion comprehension: The impact of
nonverbal intelligence. Journal of Genetic Psychology,
171(2), 101.
}
\description{
This dataset contains the correlation fitted in Albanese,
Stasio, Chiacchio, Fiorilli, and Pons (2010)
}
\details{
CPM = Colored Progression Matrices, a measure of non-verbal
intelligence. External, Mental, and Reflexive are three
stages of heirarchical development.
}
\examples{
data(albanese)
albanese
}
\references{
Fife, D.A., Rodgers, J.L., & Mendoza, J. L. (2013). Model
conditioned data elasticity in path analysis: Assessing the
"confoundability"
}
\keyword{datasets}

