\docType{data}
\name{mee}
\alias{mee}
\title{Correlation matrix for Meeuwisse, Born, and Severiens (2011)}
\format{A data frame with 7 observations on the following 7 variables.
\describe{ \item{list("SOF")}{a numeric vector} \item{list("IWF")}{a numeric
vector} \item{list("PWF")}{a numeric vector} \item{list("FSC")}{a numeric
vector} \item{list("FSF")}{a numeric vector} \item{list("EF")}{a numeric
vector} \item{list("GPA")}{a numeric vector} }}
\source{
Meeuwisse, M., Born, M. P., & Severiens, S. E. (2011). The
family-study interface and academic outcomes: Testing a
structural model. Journal of Educational Psychology,
103(4), 982 - 990.
}
\description{
This dataset contains the correlation fitted in Meeuwisse,
Born, and Severiens (2011)
}
\details{
SOF = Support of Family, IWF = Involvement with Family, PWF
= Participation with Family, FSC = Family Study Conflict,
FSF = Family Study Facilitation, EF = Effort, GPA =
university GPA.
}
\examples{
data(mee)
mee
}
\references{
%% ~~ possibly secondary sources and usages ~~
}
\keyword{datasets}

