% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseAMdct}
\alias{parseAMdct}
\title{Format AM dct File for Use with DE}
\usage{
parseAMdct(file)
}
\arguments{
\item{file}{a file location from which to read the \code{dct} file}
}
\value{
a \code{data.frame} in a format suitable for use with \code{mml} as
a \code{paramTab}.
}
\description{
Takes an \code{AM dct} file and formats it for use with the \code{mml} method
as \code{paramTab}.
}
\author{
Paul Bailey
}
