\name{residuals.DE.fit}
\alias{residuals.DE.fit}
\title{Extract residuals of model}
\usage{
\method{residuals}{DE.fit}(object, ...)
}
\arguments{
  \item{object}{an \code{OpenMx} or \code{sem} model.}
}
\description{
Given a fitted model, \code{residuals} extracts the
residual matrix for easier inspection of the source of
misfit.
}

