% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RMSR}
\alias{RMSR}
\title{Compute RMSR}
\usage{
RMSR(data, model)
}
\arguments{
\item{data}{a correlation matrix}

\item{model}{an mxmodel object}
}
\value{
the RMSR of the model
}
\description{
Compute Root Mean Square Residual (RMSR)
}
\author{
Dustin Fife
}

